channge service,exception and ,controler folders

+++ src/main/java/com/begginers/sim/survey/repository/SurveyRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SurveyRepository extends JpaRepository<Survey, Short> {
+public interface SurveyRepository extends JpaRepository<Survey, Long> {
     // ToDo
 //  Crete a custom method, custom query
 }
